Correct MIDI reresentation of tied notes
We used simply to emit the note and then emit a rest for the remainder of the tie.  But of course the player sounds the notes for the correct duration, so this rest is audible as a pause.

We now store the tied note and emit the note when we process the remainder of the tie, adding the duration of both parts.