Change representation of section variants
If a section has a variant ending, there may be one or more variant
numbers that all point to the same bar number.  I had originally
represented this as an array which was stupid bacause it required
variant numbers to be contiguous.  I now use a map from variant number
to (the same) bar number - wasteful but effective.